Reference

Pasteee.Section — Type

Section(contents; name = "", syntax = "")

Creates a paste Section with contents. The syntax argument determines syntax highlight.

Pasteee.delete — Method

delete(appkey, id)

Deletes paste id from Paste.ee.

Pasteee.get — Method

get(appkey, id)

Fetch paste id from Paste.ee.

Pasteee.paste — Method

paste(appkey, sections; description = "", expiration = "never")

Submit a paste to Paste.ee and return its id. The sections argument is either a Vector of Section objects, or a single Section object.

The expiration setting can be set to "never" (default), or to a number of seconds given as a String (e.g., "3600" for one hour).

See https://pastee.github.io/docs/.

Pasteee.paste — Method

paste(appkey, contents::String; name = "", syntax = "", ...)

Paste a single section with contents, name and syntax given.

Pasteee.pastes — Method

pastes(appkey; perpage = 25, page = 1)

Retrieve all pastes, organized in pages containing perpage entries. Returns entries in page number page.